Piney/requirements.txt
openai==1.2.2
pinecone-client==2.2.1
python-dotenv